Move pubkey cache to its own file
