+++ packages/beacon-state-transition/src/allForks/util/index.ts
@@ -7,4 +7,5 @@ export * from "./epochShuffling";
 export * from "./epochStakeSummary";
 export * from "./flat";
 export * from "./indexedSyncCommittee";
+export * from "./pubkeyCache";
 export * from "./weakSubjectivity";
